feat: refactor SplashScreen to use custom useSplash hook for improved state management
